Cast expectedLength to ssize_t where it is used as a signed number. Fixes buffer overrun memory corruption segfault.

+++ src/AsynDriverInterface.cc
@@ -906,7 +906,7 @@ readHandler()
     size_t bytesToRead = peeksize;
     size_t buffersize;
 
-    if (expectedLength > 0)
+    if ((ssize_t) expectedLength > 0)
     {
         buffersize = expectedLength;
         if (peeksize > 1)
